PDF Reader/requirement.txt
faiss-cpu
PyPDF2
langchain_google_genai
streamlit
langchain_community
streamlit